Remove 'San Antonio Symphony' strike
As of June 16th: "With deep regret, the Board of Directors of the Symphony Society of San Antonio announces the dissolution of the San Antonio Symphony.  By unanimous vote, the Board has initiated the requisite steps to file for Chapter 7 bankruptcy.  The assets of the Symphony now lie in the hands of a Trustee who will liquidate them, pay what creditors remain, and close the doors." (from https://sasymphony.org/)